# fitness-cms/web-cms/public/swagger.yaml
openapi: 3.0.3
info:
  title: FitToday CMS API
  description: |
    API for the FitToday fitness CMS platform. Connects personal trainers with students
    through programs, workouts, and real-time communication.

    ## Authentication
    All protected endpoints require a Firebase Auth token in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```

    ## Roles
    - **trainer** - Personal trainers who create programs and manage students
    - **student** - Students who follow programs and track progress
    - **admin** - Platform administrators

    ## Plans
    | Plan    | Price (BRL) | Commission | Max Programs | Max Students |
    |---------|-------------|------------|--------------|--------------|
    | Starter | Free        | 10%        | 3            | 5            |
    | Pro     | R$97/mo     | 5%         | Unlimited    | Unlimited    |
    | Elite   | R$197/mo    | 0%         | Unlimited    | Unlimited    |
  version: 1.0.0
  contact:
    name: FitToday Engineering
servers:
  - url: https://fittoday.vercel.app
    description: Production
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase Authentication ID token
  schemas:
...
    Program:
      type: object
      properties:
        id:
          type: string
        trainerId:
          type: string
        studentId:
          type: string
        title:
          type: string
        description:
          type: string
        coverImageURL:
          type: string
        previewVideoURL:
          type: string
        workoutPdfUrl:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        duration:
          type: object
          properties:
            weeks:
              type: integer
            daysPerWeek:
              type: integer
            avgSessionMinutes:
              type: integer
        requirements:
          type: object
          properties:
            equipment:
              type: array
              items:
                type: string
            fitnessLevel:
              type: string
            prerequisites:
              type: string
        content:
          type: object
          properties:
            totalWorkouts:
              type: integer
            totalExercises:
              type: integer
            includesNutrition:
              type: boolean
            includesSupplements:
              type: boolean
            hasVideoGuides:
              type: boolean
        stats:
          type: object
          properties:
            totalSales:
              type: integer
            activeStudents:
              type: integer
            completionRate:
              type: number
            averageRating:
              type: number
            totalReviews:
              type: integer
        weeks:
          type: array
          items:
            type: object
        status:
          type: string
          enum: [draft, published, archived]
        visibility:
          type: string
          enum: [public, private]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
    CreateProgramRequest:
      type: object
      required:
        - title
        - description
      properties:
        title:
          type: string
        description:
          type: string
        studentId:
          type: string
        coverImageURL:
          type: string
        previewVideoURL:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
          default: intermediate
        duration:
          type: object
          properties:
            weeks:
              type: integer
              default: 8
            daysPerWeek:
              type: integer
              default: 4
            avgSessionMinutes:
              type: integer
              default: 60
        requirements:
          type: object
          properties:
            equipment:
              type: array
              items:
                type: string
            prerequisites:
              type: string
        content:
          type: object
          properties:
            totalWorkouts:
              type: integer
            totalExercises:
              type: integer
            includesNutrition:
              type: boolean
            includesSupplements:
              type: boolean
            hasVideoGuides:
              type: boolean
        weeks:
          type: array
          items:
            type: object
        status:
          type: string
          enum: [draft, published]
          default: draft
        visibility:
          type: string
          enum: [public, private]
          default: public

    # ---- Exercises ----
...
    StripeConnectRequest:
      type: object
      required:
        - trainerId
        - email
      properties:
        trainerId:
          type: string
        email:
          type: string
        existingAccountId:
          type: string
        returnUrl:
          type: string
    StripeAccountStatus:
      type: object
      properties:
        connected:
          type: boolean
        onboardingComplete:
          type: boolean
        chargesEnabled:
          type: boolean
        payoutsEnabled:
          type: boolean
        accountId:
          type: string
        email:
          type: string
        detailsSubmitted:
          type: boolean
...
paths:
...
  /api/programs:
    get:
      tags: [Programs]
      summary: List programs
      security: []
      description: |
        If trainerId is provided, lists that trainer's programs (any status).
        Without trainerId, returns only published+public programs (marketplace).
      parameters:
        - name: trainerId
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
        - name: visibility
          in: query
          schema:
            type: string
            enum: [public, private]
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Program list
          content:
            application/json:
              schema:
                type: object
                properties:
                  programs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Program"
                  total:
                    type: integer

    post:
      tags: [Programs]
      summary: Create a program
      description: |
        Trainer-only. Supports multipart/form-data for cover image, PDF, and video uploads.
        Validates program limit based on trainer's plan (Starter: max 3).
      security:
        - BearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: JSON string of CreateProgramRequest
                cover:
                  type: string
                  format: binary
                  description: Cover image (JPEG/PNG/WebP, max 10MB)
                pdf:
                  type: string
                  format: binary
                  description: Workout PDF (max 10MB)
                video:
                  type: string
                  format: binary
                  description: Preview video (max 100MB)
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramRequest"
      responses:
        "201":
          description: Program created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  coverImageURL:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
        "400":
          description: Missing required fields or invalid file
        "401":
          description: Unauthorized
        "403":
          description: Program limit reached (upgrade plan)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  code:
                    type: string
                    enum: [PROGRAM_LIMIT_REACHED]
  /api/programs/{id}:
    get:
      tags: [Programs]
      summary: Get a program
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Program details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: Program not found

    put:
      tags: [Programs]
      summary: Update a program
      description: Trainer-only. Updates only provided fields. Supports multipart for file uploads.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: JSON string with fields to update
                cover:
                  type: string
                  format: binary
                pdf:
                  type: string
                  format: binary
                video:
                  type: string
                  format: binary
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramRequest"
      responses:
        "200":
          description: Program updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  coverImageURL:
                    type: string
                  updatedAt:
                    type: string
                    format: date-time
        "401":
          description: Unauthorized
        "403":
          description: Not the program owner
        "404":
          description: Program not found

    delete:
      tags: [Programs]
      summary: Archive a program (soft delete)
      description: |
        Trainer-only. Sets status to "archived" and visibility to "private".
        Cannot delete programs with active students.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Program archived
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                    enum: [archived]
        "400":
          description: Cannot delete with active students
        "401":
          description: Unauthorized
        "403":
          description: Not the program owner
        "404":
          description: Program not found
...
  /api/stripe/connect:
    post:
      tags: [Stripe]
      summary: Create Stripe Connect account for trainer
      security: []
      description: Creates a Stripe Connect Express account and returns the onboarding URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StripeConnectRequest"
      responses:
        "200":
          description: Account created with onboarding URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    description: Stripe onboarding URL
                  accountId:
                    type: string
        "400":
          description: Missing required fields

    get:
      tags: [Stripe]
      summary: Get Stripe Connect account status
      security: []
      parameters:
        - name: accountId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Account status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeAccountStatus"
...